+++ geometry/geometry.h
@@ -6,28 +6,28 @@
 typedef struct vector vector;
 struct vector
 {
-	double xi;
-	double yj;
-	double zk;
+	float xi;
+	float yj;
+	float zk;
 };
 
 typedef struct quaternion quaternion;
 struct quaternion
 {
 	// scalar component
-	double sc;
+	float sc;
 
 	// i, j, and k component
-	double xi;
-	double yj;
-	double zk;
+	float xi;
+	float yj;
+	float zk;
 };
 
 typedef struct quat_raw quat_raw;
 struct quat_raw
 {
 	// the rotation amount in angle degrees
-	double theta;
+	float theta;
 
 	// vector about which rotation is done
 	vector vectr;
@@ -42,16 +42,16 @@ void sum(vector* C, vector* A, vector* B);
 void diff(vector* C, vector* A, vector* B);
 
 // c = A * sc
-void multiply_scalar(vector* C, vector* A, double sc);
+void multiply_scalar(vector* C, vector* A, float sc);
 
 // C = A X B
 void cross(vector* C, vector* A, vector* B);
 
 // A.B
-double dot(vector* A, vector* B);
+float dot(vector* A, vector* B);
 
 // angle between vectors in degrees
-double angle_between_vectors(vector* A, vector* B);
+float angle_between_vectors(vector* A, vector* B);
 
 // finding unit vector fails for the vector (0,0,0), hence will return -1 else returns 0
 int unit_vector(vector* unitResult, vector* A);
@@ -63,10 +63,10 @@ void parallel_component(vector* C, vector* A, vector* B);
 void perpendicular_component(vector* C, vector* A, vector* B);
 
 // get magnitude of the vector
-double magnitude_vector(vector* D);
+float magnitude_vector(vector* D);
 
 // get magnitude of the quaternion
-double norm(quaternion* D);
+float norm(quaternion* D);
 
 // multiply quaternions with hamilton product in order as C = AB
 void hamilton_product(quaternion* C, quaternion* A, quaternion* B);
@@ -80,13 +80,13 @@ void conjugate(quaternion* destination, quaternion* source);
 void reciprocal(quaternion* destination, quaternion* source);
 
 // rotates a given vector by a given quaternion, must return 0
-double rotate_vector(vector* result, quaternion* rotation_quaternion, vector* initial);
+float rotate_vector(vector* result, quaternion* rotation_quaternion, vector* initial);
 
 // averaged update
-void update_vector(vector* result, vector* updated_value, double factor);
+void update_vector(vector* result, vector* updated_value, float factor);
 
 // Spherical linear interpolation of two quaternions
-void slerp_quaternion(quaternion* Result, quaternion* A, double factorA, quaternion* B);
+void slerp_quaternion(quaternion* Result, quaternion* A, float factorA, quaternion* B);
 
 // this method returns the quaternion rotation that caused tips of Ai and Bi non collinear vectors
 // of the object, to move to final positions Af and Bf respectively
