+++ gy86/hmc.c
@@ -58,9 +58,9 @@ int get_scaled_HMCdata(HMCdatascaled* result)
     data.magnz = (data.magnz << 8) | ((data.magnz >> 8) & 0x00ff);
 
     // in mG, milli Gauss
-    result->magn.xi = ((((double)(data.magnx)) * 0.92) - offsets.magn.xi)/1166.56;
-    result->magn.yj = ((((double)(data.magny)) * 0.92) - offsets.magn.yj)/1219.04;
-    result->magn.zk = ((((double)(data.magnz)) * 0.92) - offsets.magn.zk)/997.28;
+    result->magn.xi = ((((float)(data.magnx)) * 0.92) - offsets.magn.xi)/1166.56;
+    result->magn.yj = ((((float)(data.magny)) * 0.92) - offsets.magn.yj)/1219.04;
+    result->magn.zk = ((((float)(data.magnz)) * 0.92) - offsets.magn.zk)/997.28;
 
     return 1;
 }
